Do not clone request and response before parsing
The documentation for Request() and Response() indicate
that you will get a TypeError if the body has already
been used. Since the clone() calls in finish() happen at
the very end of the processing, the request body will most
likely have been used, and the error is thrown. Better to
just let the supplied parse() function do the clone() if
it is appropriate.